+++ src/h5core/h5t_errorhandling.c
@@ -16,13 +16,13 @@ h5tpriv_error_local_elem_nexist (
 	char s[1024];
 
 	int num_chars_printed = snprintf (s, sizeof(s), "%lld,",
-					  local_vertex_indices[0]);
+					  (long long)local_vertex_indices[0]);
 	int i;
 	int num_vertices = t->ref_element->num_faces[0];
 	for (i = 1; i < num_vertices; i++) {
 		num_chars_printed += snprintf (
 			s + num_chars_printed, sizeof (s) - num_chars_printed,
-			"%lld,", local_vertex_indices[i]);
+			"%lld,", (long long)local_vertex_indices[i]);
 		if ((sizeof (s) - num_chars_printed) < 32) {
 			// buffer to small
 			return h5_error_internal (f, __FILE__, __func__, __LINE__);
