+++ src/h5core/h5t_errorhandling_private.h
@@ -24,14 +24,14 @@ h5tpriv_error_local_elem_nexist (
 	h5_error(				\
 		f,				\
 		H5_ERR_INVAL,			\
-		"Level %lld doesn't exist.", level_id );
+		"Level %lld doesn't exist.", (long long)level_id );
 
 #define h5tpriv_error_global_id_nexist( f, name, id )		\
 	h5_error(						\
 		f,						\
 		H5_ERR_NOENTRY,					\
 		"%s with global id %lld does not exist!",	\
-		name, id );
+		name, (long long)id );
 
 
 #define h5tpriv_error_global_triangle_id_nexist( f, vids )			\
@@ -39,14 +39,14 @@ h5tpriv_error_local_elem_nexist (
 		f,							\
 		H5_ERR_NOENTRY,						\
 		"Triangle with global vertex ids (%lld,%lld,%lld) doesn't exist!", \
-		vids[0], vids[1], vids[2] );
+		(long long)vids[0], (long long)vids[1], (long long)vids[2] );
 
 #define h5tpriv_error_local_triangle_nexist( f, indices )	\
 	h5_error(							\
 		f,							\
 		H5_ERR_NOENTRY,						\
 		"Triangle with global vertex ids (%lld,%lld,%lld) doesn't exist!", \
-		indices[0], indices[1], indices[2] );
+		(long long)indices[0], (long long)indices[1], (long long)indices[2] );
 
 
 #define h5tpriv_error_store_boundaryface_local_id( f, local_fid )	     \
@@ -54,6 +54,6 @@ h5tpriv_error_local_elem_nexist (
 		f,						     \
 		H5_ERR_INVAL,					     \
 		"Boundary face with local id %lld is not on level 0!", \
-		local_fid );
+		(long long)local_fid );
 
 #endif
